Renaming and lint cleanup.
