+++ app/api/config/semanticSearch.ts
@@ -1,2 +1,4 @@
+/** @format */
+
 const { SEMANTIC_SEARCH_URL } = process.env;
 export default SEMANTIC_SEARCH_URL || 'http://localhost:5000/semanticSearch/searchOneDoc';
